// src/common/enums/common-response.enums.ts
export enum CommonResponseCodes {
  VersionInfoFetch = 'VersionInfoFetch',
  HealthCheck = 'HealthCheck',
  EncryptSuccess = 'EncryptSuccess',
  EncryptFailed = 'EncryptFailed',
}

export enum CommonResponseMessages {
  VersionFetched = 'Version fetched successfully',
  HealthOk = 'Health check OK',
  EncryptSuccess = 'Data encrypted successfully',
  EncryptFailed = 'Data encryption failed',
}